Quantity 0 Bug Fixed in CartActivity
There was a small issue with the Cart Activity displaying food items in
the cart even when there quantity was updated to zero(removed from the
cart by the user). This has been corrected.
The node of the item that is removed from the cart is also deleted from
the temp database.